+++ .github/workflows/release.yml
@@ -88,6 +88,7 @@ jobs:
           - cni-calico-deep
           - deep
           - viz
+          - tracing
           - default-policy-deny
           - external
           - rsa-ca
