+++ test/integration/tracing/tracing/testdata/emojivoto.yaml
@@ -1,4 +1,9 @@
 apiVersion: v1
+kind: Namespace
+metadata:
+  name: tracing-test
+---
+apiVersion: v1
 kind: ServiceAccount
 metadata:
   name: emoji
