+++ test/integration/tracing/tracing/testdata/jaeger-aio-values.yaml
@@ -1,0 +1,12 @@
+provisionDataStore:
+  cassandra: false
+allInOne:
+  enabled: true
+storage:
+  type: memory
+agent:
+  enabled: false
+collector:
+  enabled: false
+query:
+  enabled: false
