+++ test/integration/tracing/tracing/testdata/otel-values.yaml
@@ -1,0 +1,18 @@
+mode: deployment
+image:
+  repository: ghcr.io/open-telemetry/opentelemetry-collector-releases/opentelemetry-collector-k8s
+command:
+  name: otelcol-k8s
+config:
+  exporters:
+    debug: { }
+    otlp/jaeger:
+      endpoint: http://jaeger-collector.tracing:4317
+      tls:
+        insecure: true
+  service:
+    pipelines:
+      traces:
+        exporters:
+          - debug
+          - otlp/jaeger
