+++ test/integration/tracing/tracing/testdata/tracing.yaml
@@ -1,0 +1,6 @@
+apiVersion: v1
+kind: Namespace
+metadata:
+  name: tracing
+  annotations:
+    linkerd.io/inject: enabled
