--- test/integration/viz/tracing/testdata/check.jaeger.golden
@@ -1,12 +1,0 @@
-linkerd-jaeger
---------------
-√ linkerd-jaeger extension Namespace exists
-√ jaeger extension pods are injected
-√ jaeger injector pods are running
-√ jaeger extension proxies are healthy
-‼ jaeger extension proxies are up-to-date
-    {{.ProxyVersionErr}}
-    see {{.HintURL}}l5d-jaeger-proxy-cp-version for hints
-√ jaeger extension proxies and cli versions match
-
-Status check results are √
